+++ DnDBot.Application/Models/Ficha/Alinhamento.cs
@@ -4,21 +4,7 @@
     /// Representa um alinhamento moral e ético para um personagem.
     /// Exemplos incluem "Leal e Bom", "Caótico e Mal", etc.
     /// </summary>
-    public class Alinhamento
+    public class Alinhamento : EntidadeBase
     {
-        /// <summary>
-        /// Identificador único do alinhamento.
-        /// </summary>
-        public string Id { get; set; }
-
-        /// <summary>
-        /// Nome do alinhamento, como "Leal e Bom (LG)".
-        /// </summary>
-        public string Nome { get; set; }
-
-        /// <summary>
-        /// Descrição detalhada do significado e características do alinhamento.
-        /// </summary>
-        public string Descricao { get; set; }
     }
 }